requirements.txt
# requirements.txt
Django==4.2
redis==5.0.0
django-redis==5.2.0
djangorestframework==3.14.0